+++ NEWS.md
@@ -1,4 +1,4 @@
-# simmer 3.6.2.9000
+# simmer 3.6.3
 
 ## New features: